+++ saleor/csv/tasks.py
@@ -12,7 +12,7 @@ from ..core.tasks import RestrictWriterDBTask
 from . import events
 from .models import ExportEvent, ExportFile
 from .notifications import send_export_failed_info
-from .utils.export import export_gift_cards, export_products, export_voucher_codes
+from .utils.export import export_products
 
 task_logger = get_task_logger(__name__)
 
@@ -21,8 +21,6 @@ class ExportTask(RestrictWriterDBTask):
     # should be updated when new export task is added
     TASK_NAME_TO_DATA_TYPE_MAPPING = {
         "export-products": "products",
-        "export-gift-cards": "gift cards",
-        "export-voucher-codes": "voucher codes",
     }
 
     def on_failure(self, exc, task_id, args, kwargs, einfo):
@@ -72,38 +70,6 @@ def export_products_task(
             pk=export_file_id
         )
     export_products(export_file, scope, export_info, file_type, delimiter)
-
-
-@app.task(name="export-gift-cards", base=ExportTask)
-def export_gift_cards_task(
-    export_file_id: int,
-    scope: dict[str, str | dict],
-    file_type: str,
-    delimiter: str = ",",
-):
-    with allow_writer():
-        # Read the ExportFile instance from the writer instead of replica as it might
-        # not be replicated yet.
-        export_file = ExportFile.objects.select_related("app", "user").get(
-            pk=export_file_id
-        )
-    export_gift_cards(export_file, scope, file_type, delimiter)
-
-
-@app.task(name="export-voucher-codes", base=ExportTask)
-def export_voucher_codes_task(
-    export_file_id: int,
-    file_type: str,
-    voucher_id: int | None,
-    ids: list[int],
-):
-    with allow_writer():
-        # Read the ExportFile instance from the writer instead of replica as it might
-        # not be replicated yet.
-        export_file = ExportFile.objects.select_related("app", "user").get(
-            pk=export_file_id
-        )
-    export_voucher_codes(export_file, file_type, voucher_id, ids)
 
 
 @app.task