+++ saleor/csv/utils/export.py
@@ -9,8 +9,6 @@ from django.utils import timezone
 
 from ...core.db.connection import allow_writer
 from ...core.utils.batches import queryset_in_batches
-from ...discount.models import VoucherCode
-from ...giftcard.models import GiftCard
 from ...product.models import Product
 from .. import FileTypes
 from ..notifications import send_export_download_link_notification
@@ -59,71 +57,6 @@ def export_products(
     save_csv_file_in_export_file(export_file, temporary_file, file_name)
     temporary_file.close()
     send_export_download_link_notification(export_file, "products")
-
-
-def export_gift_cards(
-    export_file: "ExportFile",
-    scope: dict[str, str | dict],
-    file_type: str,
-    delimiter: str = ",",
-):
-    from ...graphql.giftcard.filters import GiftCardFilter
-
-    file_name = get_filename("gift_card", file_type)
-
-    queryset = get_queryset(GiftCard, GiftCardFilter, scope)
-    # only unused gift cards codes can be exported
-    queryset = queryset.filter(used_by_email__isnull=True)
-
-    export_fields = ["code"]
-    temporary_file = create_file_with_headers(export_fields, delimiter, file_type)
-
-    export_gift_cards_in_batches(
-        queryset,
-        export_fields,
-        delimiter,
-        temporary_file,
-        file_type,
-    )
-
-    save_csv_file_in_export_file(export_file, temporary_file, file_name)
-    temporary_file.close()
-    send_export_download_link_notification(export_file, "gift cards")
-
-
-def export_voucher_codes(
-    export_file: "ExportFile",
-    file_type: str,
-    voucher_id: int | None = None,
-    ids: list[int] | None = None,
-    delimiter: str = ",",
-):
-    file_name = get_filename("voucher_code", file_type)
-
-    qs = VoucherCode.objects.using(settings.DATABASE_CONNECTION_REPLICA_NAME).all()
-    if voucher_id:
-        qs = VoucherCode.objects.using(
-            settings.DATABASE_CONNECTION_REPLICA_NAME
-        ).filter(voucher_id=voucher_id)
-    if ids:
-        qs = VoucherCode.objects.using(
-            settings.DATABASE_CONNECTION_REPLICA_NAME
-        ).filter(id__in=ids)
-
-    export_fields = ["code"]
-    temporary_file = create_file_with_headers(export_fields, delimiter, file_type)
-
-    export_voucher_codes_in_batches(
-        qs,
-        export_fields,
-        delimiter,
-        temporary_file,
-        file_type,
-    )
-
-    save_csv_file_in_export_file(export_file, temporary_file, file_name)
-    temporary_file.close()
-    send_export_download_link_notification(export_file, "voucher codes")
 
 
 def get_filename(model_name: str, file_type: str) -> str:
@@ -220,40 +153,6 @@ def export_products_in_batches(
         )
 
         append_to_file(export_data, headers, temporary_file, file_type, delimiter)
-
-
-def export_gift_cards_in_batches(
-    queryset: "QuerySet",
-    export_fields: list[str],
-    delimiter: str,
-    temporary_file: Any,
-    file_type: str,
-):
-    for batch_pks in queryset_in_batches(queryset, BATCH_SIZE):
-        gift_card_batch = GiftCard.objects.using(
-            settings.DATABASE_CONNECTION_REPLICA_NAME
-        ).filter(pk__in=batch_pks)
-
-        export_data = list(gift_card_batch.values(*export_fields))
-
-        append_to_file(export_data, export_fields, temporary_file, file_type, delimiter)
-
-
-def export_voucher_codes_in_batches(
-    queryset: "QuerySet",
-    export_fields: list[str],
-    delimiter: str,
-    temporary_file: Any,
-    file_type: str,
-):
-    for batch_pks in queryset_in_batches(queryset, BATCH_SIZE):
-        voucher_codes_batch = VoucherCode.objects.using(
-            settings.DATABASE_CONNECTION_REPLICA_NAME
-        ).filter(pk__in=batch_pks)
-
-        export_data = list(voucher_codes_batch.values(*export_fields))
-
-        append_to_file(export_data, export_fields, temporary_file, file_type, delimiter)
 
 
 def append_to_file(