+++ saleor/graphql/csv/schema.py
@@ -7,7 +7,7 @@ from ..core.descriptions import DEPRECATED_EXPORT_MUTATIONS
 from ..core.fields import FilterConnectionField, PermissionsField
 from ..core.utils import from_global_id_or_error
 from .filters import ExportFileFilterInput
-from .mutations import ExportGiftCards, ExportProducts, ExportVoucherCodes
+from .mutations import ExportProducts
 from .resolvers import resolve_export_file, resolve_export_files
 from .sorters import ExportFileSortingInput
 from .types import ExportFile, ExportFileCountableConnection
@@ -44,11 +44,5 @@ class CsvQueries(graphene.ObjectType):
 
 class CsvMutations(graphene.ObjectType):
     export_products = ExportProducts.Field(
-        deprecation_reason=DEPRECATED_EXPORT_MUTATIONS
-    )
-    export_gift_cards = ExportGiftCards.Field(
-        deprecation_reason=DEPRECATED_EXPORT_MUTATIONS
-    )
-    export_voucher_codes = ExportVoucherCodes.Field(
         deprecation_reason=DEPRECATED_EXPORT_MUTATIONS
     )