+++ saleor/webhook/event_types.py
@@ -64,7 +64,6 @@ class WebhookEventAsyncType:
     GIFT_CARD_SENT = "gift_card_sent"
     GIFT_CARD_STATUS_CHANGED = "gift_card_status_changed"
     GIFT_CARD_METADATA_UPDATED = "gift_card_metadata_updated"
-    GIFT_CARD_EXPORT_COMPLETED = "gift_card_export_completed"
 
     MENU_CREATED = "menu_created"
     MENU_UPDATED = "menu_updated"
@@ -211,7 +210,6 @@ class WebhookEventAsyncType:
     VOUCHER_CODES_CREATED = "voucher_codes_created"
     VOUCHER_CODES_DELETED = "voucher_codes_deleted"
     VOUCHER_METADATA_UPDATED = "voucher_metadata_updated"
-    VOUCHER_CODE_EXPORT_COMPLETED = "voucher_code_export_completed"
 
     OBSERVABILITY = "observability"
 
@@ -354,10 +352,6 @@ class WebhookEventAsyncType:
         },
         GIFT_CARD_METADATA_UPDATED: {
             "name": "Gift card metadata updated",
-            "permission": GiftcardPermissions.MANAGE_GIFT_CARD,
-        },
-        GIFT_CARD_EXPORT_COMPLETED: {
-            "name": "Gift card export completed",
             "permission": GiftcardPermissions.MANAGE_GIFT_CARD,
         },
         MENU_CREATED: {
@@ -825,10 +819,6 @@ class WebhookEventAsyncType:
         },
         VOUCHER_METADATA_UPDATED: {
             "name": "Voucher metadata updated",
-            "permission": DiscountPermissions.MANAGE_DISCOUNTS,
-        },
-        VOUCHER_CODE_EXPORT_COMPLETED: {
-            "name": "Voucher code export completed",
             "permission": DiscountPermissions.MANAGE_DISCOUNTS,
         },
         OBSERVABILITY: {