+++ saleor/webhook/tests/fixtures/subscription_webhooks.py
@@ -262,14 +262,6 @@ def subscription_gift_card_metadata_updated_webhook(subscription_webhook):
 
 
 @pytest.fixture
-def subscription_gift_card_export_completed_webhook(subscription_webhook):
-    return subscription_webhook(
-        queries.GIFT_CARD_EXPORT_COMPLETED,
-        WebhookEventAsyncType.GIFT_CARD_EXPORT_COMPLETED,
-    )
-
-
-@pytest.fixture
 def subscription_menu_created_webhook(subscription_webhook):
     return subscription_webhook(
         queries.MENU_CREATED, WebhookEventAsyncType.MENU_CREATED
@@ -1263,14 +1255,6 @@ def subscription_voucher_webhook_with_meta(subscription_webhook):
 def subscription_voucher_metadata_updated_webhook(subscription_webhook):
     return subscription_webhook(
         queries.VOUCHER_METADATA_UPDATED, WebhookEventAsyncType.VOUCHER_METADATA_UPDATED
-    )
-
-
-@pytest.fixture
-def subscription_voucher_code_export_completed_webhook(subscription_webhook):
-    return subscription_webhook(
-        queries.VOUCHER_CODE_EXPORT_COMPLETED,
-        WebhookEventAsyncType.VOUCHER_CODE_EXPORT_COMPLETED,
     )